Revert to using template tag for maintenance_alerts
Resolves tests failing due to db call in context_processors.py  and simplifies return logic
