+++ corehq/util/context_processors.py
@@ -11,7 +11,6 @@ from corehq.apps.accounting.models import BillingAccount, SubscriptionType
 from corehq.apps.accounting.utils import domain_has_privilege
 from corehq.apps.analytics.utils.hubspot import is_hubspot_js_allowed_for_request
 from corehq.apps.hqwebapp.utils import get_environment_friendly_name
-from corehq.apps.hqwebapp.models import MaintenanceAlert
 
 COMMCARE = 'commcare'
 COMMTRACK = 'commtrack'
@@ -232,20 +231,6 @@ def mobile_experience_hidden_by_toggle(request):
             return True
     return False
 
-
-def maintenance_alerts(request):
-    active_alerts = MaintenanceAlert.get_active_alerts()
-    context = {}
-    if active_alerts:
-        domain = getattr(request, 'domain', None)
-        context.update({
-            'maintenance_alerts': [
-                alert for alert in active_alerts
-                if (not alert.domains
-                    or domain in alert.domains)
-            ],
-        })
-    return context
 
 
 def subscription_banners(request):
